Add custom 2D Lit
