--- Assets/Game/Scripts/GameM.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Game.MainRoom;
-
-public class GameM : MonoBehaviour
-{
-    public DialogueController diC;
-    public MemoContorller memoC;
-
-    public void NextMemo()
-    {
-        memoC.NextDay();
-    }
-}
