--- Assets/Game/Scripts/GameM.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 31cc67255b848ad408a980417750ba09
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
