--- Assets/Game/Scripts/GameManager.cs
@@ -1,47 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using UnityEngine;
-
-public class GameManager : MonoBehaviour
-{
-    #region Singleton
-    private static GameManager instance;
-    public static GameManager Instance
-    {
-        get { return instance; }
-    }
-
-    private void Awake()
-    {
-        if (instance != null && instance != this)
-        {
-            Destroy(gameObject);
-        }
-        else
-        {
-            instance = this;
-            DontDestroyOnLoad(gameObject);
-        }
-    }
-    #endregion
-
-    public SceneController SceneController;
-
-    GameObject[] Controllers;
-
-    private void OnEnable()
-    {
-        Controllers = GameObject.FindGameObjectsWithTag("Controller");
-        Controllers.Select(t => t.TryGetComponent(out SceneController));
-    }
-
-    public void EXITGAME()
-    {
-#if UNITY_EDITOR
-        UnityEditor.EditorApplication.isPlaying = false;
-#else
-        Application.Quit();
-#endif
-    }
-}
