+++ Assets/Game/Shaders/Custom-Sprte-Lit-Default.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 51fe7fd37013f584e874d548a47da69f
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
