+++ UserSettings/EditorUserSettings.asset
@@ -18,10 +18,10 @@ EditorUserSettings:
       value: 5508010050025b590b5b5970127a0944464f4b7b7c7f7f632f7a1c6abbb16761
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5508010050025b590b5b5970127a0944464f4b7b7c7f7f632f7a1c6abbb16761
+      value: 0054000756020b5a5f0c5c7a49225a4441164a72797e7361292a1c30e1b66c3d
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 5209565e560c580f0c0c5d72497b5c44414f4d2f7c7f74612e784463e1b06368
+      value: 5508010050025b590b5b5970127a0944464f4b7b7c7f7f632f7a1c6abbb16761
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 0054000756020b5a5f0c5c7a49225a4441164a72797e7361292a1c30e1b66c3d
@@ -30,13 +30,7 @@ EditorUserSettings:
       value: 5a08505207025d0f5c57082144720b44464e1e7d787e77602c7f4b64b5e36d6f
       flags: 0
     RecentlyUsedSceneGuid-8:
-      value: 0054000756020b5a5f0c5c7a49225a4441164a72797e7361292a1c30e1b66c3d
-      flags: 0
-    RecentlyUsedSceneGuid-7:
-      value: 5a08505207025d0f5c57082144720b44464e1e7d787e77602c7f4b64b5e36d6f
-      flags: 0
-    RecentlyUsedSceneGuid-8:
-      value: 0054000756020b5a5f0c5c7a49225a4441164a72797e7361292a1c30e1b66c3d
+      value: 5209565e560c580f0c0c5d72497b5c44414f4d2f7c7f74612e784463e1b06368
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
